% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicator}
\alias{indicator}
\alias{Prevedere.api.indicator}
\alias{Prevedere.api.indicator_series}
\title{Access Prevedere indicators}
\usage{
Prevedere.api.indicator(provider, provider_id)

Prevedere.api.indicator_series(provider, provider_id, freq = c("Annual",
  "SemiAnnual", "Quarterly", "Monthly", "BiWeekly", "Weekly", "Daily"),
  calculation = c("None", "PeriodOverPeriod", "YearOverYear",
  "ThreePeriodMoving", "FivePeriodMoving", "ThreePeriodYearOverYear"),
  start_date = NULL, end_date = NULL, offset_periods = 0,
  raw = FALSE)
}
\arguments{
\item{provider}{Code for a data provider, can be hexidecimal or abbreviated
name.}

\item{provider_id}{Specific ProviderID for the indicator.}

\item{freq}{Frequency of indicator to retrieve. Must be one of "Annual",
"SemiAnnual", "Quarterly", "Monthly", "BiWeekly", "Weekly", or "Daily".}

\item{calculation}{Calculation to transform the indicator. Must be one of
"None", "PeriodOverPeriod", "YearOverYear", "ThreePeriodMoving",
"FivePeriodMoving", or "ThreePeriodYearOverYear".}

\item{start_date, end_date}{Start and end dates for the indicator. Each should be
either a date or a character string capable of being coerced to a date.
Setting a date to \code{NULL} will result in the historical data being unbounded
in that direction.}

\item{offset_periods}{Number of periods to offset.}

\item{raw}{Logical value indicating if data should be returned in its raw form
(typically nested lists) or formatted as appropriate, usually to a dataframe.}
}
\value{
A list. Prevedere.api.indicator returns metadata for
the target indicator, while Prevedere.api.indicator_series returns the actual
data for the indicator (in addition to the metadata).
}
\description{
Access indicator data and metadata.
}
\examples{
\dontrun{

## Return indicator metadata
Prevedere.api.indicator(provider = "BLS", provider_id = "CES3133231058")

## Return indicator data
Prevedere.api.indicator_series(provider = "BLS", provider_id = "CES3133231058",
freq = "Monthly", calculation = "None", start_date = "2010-01-01",
offset_periods = 0)

## Return indicator data unformatted
Prevedere.api.indicator_series(provider = "BLS", provider_id = "CES3133231058",
freq = "Monthly", calculation = "None", start_date = "2010-01-01",
offset_periods = 0,raw = TRUE)
}

}
