% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Prevedere.api.correlation}
\alias{Prevedere.api.correlation}
\title{Correlation}
\usage{
Prevedere.api.correlation(endog_provider, endog_provider_id, exog_provider,
  exog_provider_id, freq = c("Annual", "SemiAnnual", "Quarterly",
  "Monthly", "BiWeekly", "Weekly", "Daily"), calculation = c("None",
  "PeriodOverPeriod", "YearOverYear", "ThreePeriodMoving",
  "FivePeriodMoving", "ThreePeriodYearOverYear"), raw = FALSE)
}
\arguments{
\item{endog_provider}{Code for the data provider of the endogenous indicator,
can be hexidecimal or abbreviated name.}

\item{endog_provider_id}{Specific ProviderID for the endogenous indicator.}

\item{exog_provider}{Code for the data provider of the exogneous indicator,
can be hexidecimal or abbreviated name.}

\item{exog_provider_id}{Specific ProviderID for the exogenous indicator.}

\item{freq}{Frequency of indicator to retrieve. Must be one of "Annual",
"SemiAnnual", "Quarterly", "Monthly", "BiWeekly", "Weekly", or "Daily".}

\item{calculation}{Calculation to transform the indicator. Must be one of
"None", "PeriodOverPeriod", "YearOverYear", "ThreePeriodMoving",
"FivePeriodMoving", or "ThreePeriodYearOverYear".}

\item{raw}{Logical value indicating if data should be returned in its raw form
(typically nested lists) or formatted as appropriate, usually a dataframe.}
}
\value{
Model results and metadata, as a list.
}
\description{
Calculates Pearson's r and other statistics at different offsets between an
endogenous and exogenous indicator.
}
\examples{
\dontrun{
Prevedere.api.correlation(endog_provider = "BLS", endog_provider_id = "CES3133231058",
exog_provider = "FRED", exog_provider_id = "PCU332313332313", freq = "Monthly",
calculation = "ThreePeriodMoving")
}
}
