% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider.R
\name{Prevedere.api.providers}
\alias{Prevedere.api.providers}
\title{Providers}
\usage{
Prevedere.api.providers(raw = FALSE)
}
\arguments{
\item{raw}{Logical value indicating if data should be returned in its raw form
(typically nested lists) or formatted as appropriate, usually a dataframe.}
}
\value{
A dataframe (or list if \code{raw = TRUE}) of provider metadata.
}
\description{
Retrieve a list of entities whose data is available in Prevedere
}
\examples{
\dontrun{
Prevedere.api.providers()
}
}
