% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Prevedere.api.raw_model}
\alias{Prevedere.api.raw_model}
\title{Title}
\usage{
Prevedere.api.raw_model(model_id, exclude_indicators = TRUE,
  as_of_date = NULL, raw = FALSE)
}
\arguments{
\item{raw}{}
}
\description{
Title
}
