% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{Prevedere.api.search}
\alias{Prevedere.api.search}
\title{Search Prevedere}
\usage{
Prevedere.api.search(query)
}
\arguments{
\item{query}{A query to send to Prevedere. This should be a named list.}
}
\value{
The result of the search.
}
\description{
Send an arbitrary query to the Prevedere API.
}
\seealso{
[Prevedere.api.fetch]
}
