% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbench.R
\name{Prevedere.api.workbench}
\alias{Prevedere.api.workbench}
\title{Workbench}
\usage{
Prevedere.api.workbench(workbench_id)
}
\arguments{
\item{workbench_id}{UUID for workbench.}
}
\value{
A list of workbench metadata.
}
\description{
Returns the indicators used in a workbench. Typically includes workbench
metadata and associated indicator metadata.
}
\examples{
\dontrun{
Prevedere.api.workbench("b8da829f7a1d4509ca5125e4699d6f0e")
}
}
