% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeration.R
\name{prevedere_aggregations}
\alias{prevedere_aggregations}
\title{Aggregation methods}
\usage{
prevedere_aggregations(key)
}
\arguments{
\item{key}{A Prevedere API key.}
}
\value{
A character vector of aggregation methods.
}
\description{
Retrieve a list of the aggregation methods currently supported by the
Prevedere API.
}
\examples{
\dontrun{
key <- "1235467abcdefg"
prevedere_aggregations(key)
}
}
\seealso{
Other enumeration functions: \code{\link{prevedere_calculations}},
  \code{\link{prevedere_frequencies}},
  \code{\link{prevedere_seasonalities}}
}
\concept{enumeration functions}
