% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeration.R
\name{prevedere_calculations}
\alias{prevedere_calculations}
\title{Calculations}
\usage{
prevedere_calculations(key)
}
\arguments{
\item{key}{A Prevedere API key.}
}
\value{
A character vector of calculations.
}
\description{
Retrieve a list of the calculations currently supported by the
Prevedere API.
}
\examples{
\dontrun{
key <- "1235467abcdefg"
prevedere_calculations(key)
}
}
\seealso{
Other enumeration functions: \code{\link{prevedere_aggregations}},
  \code{\link{prevedere_frequencies}},
  \code{\link{prevedere_seasonalities}}
}
\concept{enumeration functions}
