% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeration.R
\name{prevedere_frequencies}
\alias{prevedere_frequencies}
\title{Frequencies}
\usage{
prevedere_frequencies(key)
}
\arguments{
\item{key}{A Prevedere API key.}
}
\value{
A character vector of frequencies.
}
\description{
Retrieve a list of the time frequencies currently supported by the
Prevedere API.
}
\examples{
\dontrun{
key <- "1235467abcdefg"
prevedere_frequencies(key)
}
}
\seealso{
Other enumeration functions: \code{\link{prevedere_aggregations}},
  \code{\link{prevedere_calculations}},
  \code{\link{prevedere_seasonalities}}
}
\concept{enumeration functions}
