% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider.R
\name{prevedere_providers}
\alias{prevedere_providers}
\title{Providers}
\usage{
prevedere_providers(key, raw = FALSE)
}
\arguments{
\item{key}{A Prevedere API key.}

\item{raw}{Logical value indicating if data should be returned in its raw form
(typically nested lists) or formatted as appropriate, usually a dataframe.}
}
\value{
A dataframe (or list if \code{raw = TRUE}) of provider metadata.
}
\description{
Retrieve a list of entities whose data is available in Prevedere
}
\examples{
\dontrun{
k <- "1235467abcdefg"

prevedere_providers(key = k)
}
}
