% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeration.R
\name{prevedere_seasonalities}
\alias{prevedere_seasonalities}
\title{Seasonalities}
\usage{
prevedere_seasonalities(key)
}
\arguments{
\item{key}{A Prevedere API key.}
}
\value{
A character vector of seasonalities.
}
\description{
Retrieve a list of the seasonalities currently supported by the
Prevedere API.
}
\examples{
\dontrun{
key <- "1235467abcdefg"
prevedere_seasonalities(key)
}
}
\seealso{
Other enumeration functions: \code{\link{prevedere_aggregations}},
  \code{\link{prevedere_calculations}},
  \code{\link{prevedere_frequencies}}
}
\concept{enumeration functions}
