% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbench.R
\name{prevedere_workbench}
\alias{prevedere_workbench}
\title{Workbench}
\usage{
prevedere_workbench(key, workbench_id)
}
\arguments{
\item{key}{A Prevedere API key.}

\item{workbench_id}{UUID for workbench.}
}
\value{
A list of workbench metadata.
}
\description{
Returns the indicators used in a workbench. Typically includes workbench
metadata and associated indicator metadata.
}
\examples{
\dontrun{
k <- "1235467abcdefg"

prevedere_workbench(key = k, workbench_id = "b8da829f7a1d4509ca5125e4699d6f0e")
}
}
\concept{workbench functions}
