% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevederer.R
\docType{package}
\name{prevederer-package}
\alias{prevederer}
\alias{prevederer-package}
\title{Prevederer}
\description{
prevederer provides an R wrapper around the
\href{https://www.prevedere.com/}{Prevedere Software} API. The package facilitates
access to the main \href{https://api.prevedere.com/index.html}{API components}, including:
}
\details{
\itemize{
\item Enumeration (lists of acceptable values)
\item Indicator
\item ForecastModel
\item Workbench
}

Direct calls can also be made to the API using \code{\link[=prevedere_fetch]{prevedere_fetch()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wkdavis/prevederer}
  \item \url{https://api.prevedere.com/index.html}
  \item \url{https://www.prevedere.com/}
  \item Report bugs at \url{https://github.com/wkdavis/prevederer/issues}
}

}
\author{
\strong{Maintainer}: Wil Davis \email{davis.3243@osu.edu} (0000-0001-9780-2576)

Other contributors:
\itemize{
  \item Hiram Foster \email{hiramfoster.co@gmail.com} [contributor]
  \item Ted Dickinson \email{tmdickinson@gmail.com} [contributor]
}

}
