% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevederer.R
\docType{package}
\name{prevederer-package}
\alias{prevederer}
\alias{prevederer-package}
\title{Prevederer}
\description{
prevederer provides an R wrapper around the
\href{https://www.prevedere.com/}{Prevedere Software} API. The package facilitates
access to the main API components, including:
}
\details{
\itemize{
\item Indicators
\item Models (correlations, raw models, and forecasts)
\item Workbenches
}

Direct calls can also be made to the API using \link{Prevedere.api.fetch}.

\strong{Note that access to the API requires an API key}. The API key is set in
\link{options} using the convenience function \link{Prevedere.api.set_key}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.prevedere.com/}
  \item \url{https://github.com/wkdavis/prevederer}
  \item Report bugs at \url{https://github.com/wkdavis/prevederer/issues}
}

}
\author{
\strong{Maintainer}: Wil Davis \email{davis.3243@osu.edu} (0000-0001-9780-2576)

}
